imports clean up, structure refactor -> new restaurant package
