+++ src/main/java/pl/kupniewski/presenter/RestaurantPresenter.java
@@ -1,7 +1,6 @@
 package pl.kupniewski.presenter;
 
 import lombok.AllArgsConstructor;
-import org.springframework.http.HttpRequest;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
@@ -10,7 +9,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import pl.kupniewski.application.order.OrderStorage;
-import pl.kupniewski.application.deliver.api.RestaurantApi;
+import pl.kupniewski.application.restaurant.RestaurantApi;
 import pl.kupniewski.application.deliver.OrderReadyToDeliver;
 import pl.kupniewski.application.menu.Menu;
 import pl.kupniewski.application.order.Order;
